universal7885: replace needed libcamera_client.so to VNDK libs.
Our old camera HALs link libcamera_client, which isn't exposed to vendor, so we
need to instead link against what was supposed to replace libcamera_client:
android.hardware.camera.common@1.0-helper. However, this was supposed to be
statically linked into the camera HAL, so work around this by building a shared
version of it and then replacing the libcamera_client link in the camera HAL with it.

Change-Id: I3288ca83df18735d718129892017799fe13acd28